Rename directory to match namespace
