+++ lib/onfleet-ruby.rb
@@ -1,32 +1,32 @@
 # Errors
-require 'onfleet-ruby/errors/onfleet_error'
-require 'onfleet-ruby/errors/authentication_error'
-require 'onfleet-ruby/errors/invalid_request_error'
-require 'onfleet-ruby/errors/connection_error'
+require 'onfleet/errors/onfleet_error'
+require 'onfleet/errors/authentication_error'
+require 'onfleet/errors/invalid_request_error'
+require 'onfleet/errors/connection_error'
 
 # Actions
-require 'onfleet-ruby/actions/create'
-require 'onfleet-ruby/actions/find'
-require 'onfleet-ruby/actions/save'
-require 'onfleet-ruby/actions/update'
-require 'onfleet-ruby/actions/get'
-require 'onfleet-ruby/actions/list'
-require 'onfleet-ruby/actions/delete'
-require 'onfleet-ruby/actions/query_metadata'
+require 'onfleet/actions/create'
+require 'onfleet/actions/find'
+require 'onfleet/actions/save'
+require 'onfleet/actions/update'
+require 'onfleet/actions/get'
+require 'onfleet/actions/list'
+require 'onfleet/actions/delete'
+require 'onfleet/actions/query_metadata'
 
 # Resources
-require 'onfleet-ruby/http'
-require 'onfleet-ruby/onfleet_object'
-require 'onfleet-ruby/recipient'
-require 'onfleet-ruby/destination'
-require 'onfleet-ruby/address'
-require 'onfleet-ruby/barcode'
-require 'onfleet-ruby/task'
-require 'onfleet-ruby/organization'
-require 'onfleet-ruby/admin'
-require 'onfleet-ruby/team'
-require 'onfleet-ruby/vehicle'
-require 'onfleet-ruby/worker'
-require 'onfleet-ruby/hub'
-require 'onfleet-ruby/webhook'
+require 'onfleet/http'
+require 'onfleet/onfleet_object'
+require 'onfleet/recipient'
+require 'onfleet/destination'
+require 'onfleet/address'
+require 'onfleet/barcode'
+require 'onfleet/task'
+require 'onfleet/organization'
+require 'onfleet/admin'
+require 'onfleet/team'
+require 'onfleet/vehicle'
+require 'onfleet/worker'
+require 'onfleet/hub'
+require 'onfleet/webhook'
