+++ lib/onfleet/actions/delete.rb
