+++ lib/onfleet/barcode.rb
