+++ lib/onfleet/errors/authentication_error.rb
