+++ lib/onfleet/errors/connection_error.rb
