+++ lib/onfleet/vehicle.rb
